toy/my-ai-coder/robot-monitor/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
pydantic==2.9.0
python-dotenv==1.0.1
